remove 1.x deprecated features
