+++ .travis.yml
@@ -4,8 +4,6 @@ sudo: false
 
 matrix:
   include:
-    - php: 7.0
-      env: COLLECT_COVERAGE=false VALIDATE_CODING_STYLE=false RUN_PHPSTAN=false IGNORE_PLATFORMS=false
     - php: 7.1
       env: COLLECT_COVERAGE=true VALIDATE_CODING_STYLE=false RUN_PHPSTAN=false IGNORE_PLATFORMS=false
     - php: 7.2
