+++ composer.json
@@ -30,7 +30,7 @@
         "issues": "https://github.com/thephpleague/uri/issues"
     },
     "require": {
-        "php": ">=7.0.13",
+        "php": ">=7.1.3",
         "ext-mbstring" : "*",
         "ext-intl" : "*",
         "ext-fileinfo": "*",
@@ -40,7 +40,7 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.0",
-        "phpunit/phpunit" : "^6.0",
+        "phpunit/phpunit" : "^7.0",
         "phpstan/phpstan": "^0.9.2",
         "phpstan/phpstan-strict-rules": "^0.9.0",
         "phpstan/phpstan-phpunit": "^0.9.4"
@@ -76,7 +76,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "2.x-dev"
         }
     },
     "config": {
