--- src/Schemes/Ws.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * League.Uri (http://uri.thephpleague.com)
- *
- * @package    League\Uri
- * @subpackage League\Uri\Schemes
- * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
- * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    1.1.1
- * @link       https://github.com/thephpleague/uri-components
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace League\Uri\Schemes;
-
-use League\Uri;
-
-class_exists(Uri\Ws::class);
