+++ src/functions_include.php
@@ -3,5 +3,4 @@
 // Don't redefine the functions if included multiple times.
 if (!function_exists('League\Uri\create')) {
     require __DIR__.'/functions.php';
-    require __DIR__.'/Schemes/deprecated.php';
 }
